+++ src/client/kotlin/liltojustice/trueadaptivemusic/client/sound/instance/AudioFileSoundInstance.kt
@@ -13,8 +13,9 @@ import net.minecraft.sound.SoundCategory
 import net.minecraft.util.Identifier
 import java.util.concurrent.CompletableFuture
 
-class AdaptiveMusicSoundInstance(private val soundFile: SoundFile)
-    : AbstractSoundInstance(Constants.TRUEADAPTIVEMUSIC_ID, SoundCategory.MUSIC, SoundInstance.createRandom()), VolumeControlled {
+class AudioFileSoundInstance(private val soundFile: SoundFile)
+    : AbstractSoundInstance(Constants.AUDIO_FILE_STREAM_ID, SoundCategory.MUSIC, SoundInstance.createRandom()),
+    VolumeControlled {
     override fun getAudioStream(loader: SoundLoader, id: Identifier, repeatInstantly: Boolean):
             CompletableFuture<AudioStream> {
         val extension = soundFile.getExtension()
