+++ src/main/resources/assets/trueadaptivemusic/sounds.json
@@ -1,5 +1,5 @@
 {
-  "trueadaptivemusicstream":
+  "audiofilestream":
   {
     "sounds": [
       {
